# Optimize `ULID#to_s` with using native `Integer#to_s`