For a decoder part of #7

Same direction as 8ccb012fe6e0cf7664535595879aa38b50f6763f, #85

```console
$ ruby -v
ruby 3.0.1p64 (2021-04-05 revision 0fb782ee38) [x86_64-darwin20]

$ bundle exec ruby benchmark/to_s.rb
Warming up --------------------------------------
ULID#to_s_with_integer_base / Before #7
                         2.228k i/100ms
ULID#to_s / After #7     5.882k i/100ms
Calculating -------------------------------------
ULID#to_s_with_integer_base / Before #7
                         23.567k (± 9.4%) i/s -    118.084k in   5.056627s
ULID#to_s / After #7     74.358k (± 5.6%) i/s -    376.448k in   5.079996s

Comparison:
ULID#to_s / After #7:    74357.6 i/s
ULID#to_s_with_integer_base / Before #7:    23567.4 i/s - 3.16x  (± 0.00) slower
```

ref: #86, #88